move LHHeadAssembly back into lhtypes, reverts a0a78b0
